chore: Update Dockerfile to clarify internal user and group
For native Docker installations, the user and group inside
the image should match that user and group ID on the host. This
allows files written to the bind mount to maintain correct ownreship.

This is important for things like migrations, logs, and other
artifacts produced by the software and desired on the host. Bind
mounts shadow the entire directory in the image, so if the process
is run by root, a UID and GID of 0 will be set for those files.

More info here:
https://jtreminio.com/blog/running-docker-containers-as-current-host-user/